refactor : 도메인, 공통부분, UI 분리를 위한 패키지 재설계
- 패키지 재설계를 위한 클래스 분배
