+++ src/main/java/carracing/utils/CarRacingProgramPhrases.java
@@ -1,6 +1,6 @@
-package carracing;
+package carracing.utils;
 
-public class CarRacingProgramPhrase {
+public class CarRacingProgramPhrases {
 
     public static final String CAR_LIST_INPUT_PHRASE = "경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).";
     public static final String RACE_COUNT_PHRASE = "시도할 회수는 몇회인가요?";
